+++ plugins/clickhouse/src/components/panel/LogsDocuments.tsx
@@ -1,0 +1,48 @@
+import { TableComposable, TableVariant, Tbody, Th, Thead, Tr } from '@patternfly/react-table';
+import React from 'react';
+
+import { ILogsData } from '../../utils/interfaces';
+import LogsDocument from './LogsDocument';
+
+interface ILogsDocumentsProps {
+  pages: ILogsData[];
+  fields?: string[];
+  showDetails?: (details: React.ReactNode) => void;
+}
+
+const LogsDocuments: React.FunctionComponent<ILogsDocumentsProps> = ({
+  pages,
+  fields,
+  showDetails,
+}: ILogsDocumentsProps) => {
+  return (
+    <TableComposable aria-label="Logs" variant={TableVariant.compact} borders={false}>
+      <Thead>
+        <Tr>
+          <Th>Time</Th>
+          {fields && fields.length > 0 ? (
+            fields.map((selectedField, index) => <Th key={index}>{selectedField}</Th>)
+          ) : (
+            <Th>Log</Th>
+          )}
+        </Tr>
+      </Thead>
+      <Tbody>
+        {pages.map((page, pageIndex) =>
+          page.documents
+            ? page.documents.map((document, documentIndex) => (
+                <LogsDocument
+                  key={`${pageIndex}_${documentIndex}`}
+                  document={document}
+                  fields={fields}
+                  showDetails={showDetails}
+                />
+              ))
+            : null,
+        )}
+      </Tbody>
+    </TableComposable>
+  );
+};
+
+export default LogsDocuments;