+++ plugins/clickhouse/src/components/panel/Panel.tsx
@@ -1,0 +1,54 @@
+import React, { memo } from 'react';
+
+import { IPluginPanelProps, PluginOptionsMissing } from '@kobsio/plugin-core';
+import { IPanelOptions } from '../../utils/interfaces';
+import Logs from './Logs';
+
+interface IPanelProps extends IPluginPanelProps {
+  options?: IPanelOptions;
+}
+
+export const Panel: React.FunctionComponent<IPanelProps> = ({
+  name,
+  title,
+  description,
+  times,
+  options,
+  showDetails,
+}: IPanelProps) => {
+  if (
+    !options ||
+    !times ||
+    (options.type === 'logs' && (!options.queries || !Array.isArray(options.queries) || options.queries.length === 0))
+  ) {
+    return (
+      <PluginOptionsMissing
+        title={title}
+        message="Options for ClickHouse panel are missing or invalid"
+        details="The panel doesn't contain the required options to render get the ClickHouse data or the provided options are invalid."
+        documentation="https://kobs.io/plugins/elasticsearch"
+      />
+    );
+  }
+
+  if (options.type === 'logs' && options.queries) {
+    <Logs
+      name={name}
+      title={title}
+      description={description}
+      queries={options.queries}
+      times={times}
+      showDetails={showDetails}
+    />;
+  }
+
+  return null;
+};
+
+export default memo(Panel, (prevProps, nextProps) => {
+  if (JSON.stringify(prevProps) === JSON.stringify(nextProps)) {
+    return true;
+  }
+
+  return false;
+});